Create first part of possible algorithm.
Added an algorithm that detects if the branch has at least one of a single player's color and adds it to a list of possible captures if it does along with a bool array to represent which branches are blank vs. which are the capturing branch.  Does not add to the list if the square has both colors attached to it nor if it has neither.